+++ ChoGGi's Library/Code/ListChoice.lua
@@ -92,7 +92,7 @@ function ChoGGi_ListChoiceDlg:Init(parent, context)
 	self.items = self.list.items
 
 	self.obj = context.obj or self.items[1].obj
-	self.custom_func = self.list.custom_func
+	self.custom_func = self.list.custom_func or self.list.callback
 	self.custom_type = self.list.custom_type
 	self.close_func = self.list.close_func
 	self.title = self.list.title