+++ Expanded Cheat Menu/Changelog.md
@@ -6,6 +6,7 @@
 ## v12.5 (Unreleased)
 ### Added:
 - ECM>Grid Info: List objects in grids (air, electricity, and water).
+- Debug>Building Path Markers: Show inside waypoints colonists take to move around.
 
 ### Changed:
 - Map Exploration will no longer save the DeepScanAvailable setting.
@@ -23,6 +24,8 @@
 - > Added some more options to BBox Toggle.
 - > Instead of changing focus to an already opened examine dialog, it'll now flash the titlebar red.
 - > Added tooltips to "hyperlinks".
+- > Added Object>Entity Surfaces: Shows list of surfaces for the object entity.
+- > Removes any placed bbox, hex shapes, and spot names when the dialog is closed.
 
 ### Fixed:
 - Delete Object didn't work with filled Mech Depots (thanks FirstGeekDanny).